editing calcs guess what
I love God! You are the BEST! Thank you so much Hashem for giving me the time to work on this things. I don't know what im doing at all but I trust that it is for soe grater purpose. otherwise Id wtf the point of all of this back and forth is. more interations more iterations. refinement. until perfection is met